update logic upgrade bonus

+++ Assets/Dark/Scripts/InGame/Upgrade/NodeLogics/NodeBonusDash.cs
@@ -23,7 +23,7 @@ namespace InGame.Upgrade
                     bonusInfo.dashSizePlus += value[level - 1];
                     break;
                 case BonusType.Damage:
-                    bonusInfo.dashDamagePlus += value[level - 1];
+                    bonusInfo.dashDamagePlus += (int)value[level - 1];
                     break;
             }
         }
